pyproject.toml
[tool.poetry]
name = "postcodes"
version = "0.1.0"
description = ""
authors = ["Eddie O'Donnell <eddieod88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.3"
numpy = "^1.23.1"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
ipykernel = "^6.15.1"
black = "^22.6.0"
pyarrow = "^9.0.0"
loguru = "^0.6.0"
notebook = "^6.4.12"
scikit-learn = "^1.1.2"
geopandas = "^0.12.1"
requests = "^2.28.1"
aiohttp = "^3.8.3"
geopy = "^2.2.0"
openpyxl = "^3.0.10"
utm = "^0.7.0"
polars = "^0.15.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
